requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
openpyxl==3.1.2
pytz==2023.3
Werkzeug==2.3.7
gunicorn==21.2.0